Sample input and output
Sample Input:-
Enter no. of rows :: 3
Enter no. of cols :: 3
Enter values to the matrix ::
Enter a[0][0] value :: 1
Enter a[0][1] value :: 2
Enter a[0][2] value :: 3
Enter a[1][0] value :: 4
Enter a[1][1] value :: 5
Enter a[1][2] value :: 6
Enter a[2][0] value :: 7
Enter a[2][1] value :: 8
Enter a[2][2] value :: 9
Sample output-
The Diagonals elements of a matrix are ::
        1
                5
                        9